+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/typeresolution/typeinference/Variable.java
@@ -1,0 +1,8 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.java.typeresolution.typeinference;
+
+public class Variable {
+}