+++ pmd-java/src/test/java/net/sourceforge/pmd/typeresolution/ClassTypeResolverTest.java
@@ -18,6 +18,7 @@ import java.util.StringTokenizer;
 
 import org.apache.commons.io.IOUtils;
 import org.jaxen.JaxenException;
+
 import org.junit.Assert;
 import org.junit.Test;